chore: rename deps/hashes.ts to deps/blake2b
